requirements.txt
Pillow>=9.0
pillow-heif>=0.12
opencv-python>=4.9
numpy>=1.24
ultralytics
openai
python-dotenv